Add http-equiv meta tag.
This is a little gross; the tag is forcibly added by nokogiri now: https://github.com/sparklemotion/nokogiri/issues/1008

If it’s going to be forcibly added, at least it should be indented properly. :'(

Let’s add it to the layout for now, but if we can remove it in the future and just rely on the modern standard (<meta charset="utf-8">), let’s do it!